some edits
I dont understand the code i wrote rn
